+++ src/components/Login/Login.jsx
@@ -1,7 +1,14 @@
 import React from 'react';
+import { useAuth0 } from '@auth0/auth0-react';
+import LogoutButton from './buttons/LogoutButton';
+import LoginButton from './buttons/LoginButton';
 
 const Login = () => {
-  return <h1>Login</h1>;
+  const { isAuthenticated } = useAuth0();
+
+  return (
+    isAuthenticated ? <LogoutButton/> : <LoginButton/>
+  );
 };
 
 export default Login;
