+++ src/components/Login/Profile.jsx
@@ -1,0 +1,18 @@
+import React from 'react';
+import { useAuth0 } from '@auth0/auth0-react';
+
+const Profile = () => {
+  const { user, isAuthenticated } = useAuth0();
+
+  return (
+    isAuthenticated && (
+      <div>
+        <img src={user.picture} alt={user.name} />
+        <h2>{user.name}</h2>
+        <p>{user.email}</p>
+      </div>
+    )
+  );
+};
+
+export default Profile;
