+++ src/components/Pages/Splash.jsx
@@ -1,40 +1,12 @@
 import React from 'react';
-import { useAuth0 } from '@auth0/auth0-react';
-
-const LoginButton = () => {
-  const { loginWithRedirect } = useAuth0();
-
-  return <button onClick={() => loginWithRedirect()}>Log In</button>;
-};
-
-const LogoutButton = () => {
-  const { logout } = useAuth0();
-
-  return <button onClick={() => logout({ returnTo: window.location.origin })}>
-    Log Out
-  </button>;
-};
-
-const Profile = () => {
-  const { user, isAuthenticated } = useAuth0();
-
-  return (
-    isAuthenticated && (
-      <div>
-        <img src={user.picture} alt={user.name} />
-        <h2>{user.name}</h2>
-        <p>{user.email}</p>
-      </div>
-    )
-  );
-};
+import Profile from '../Login/Profile';
+import Login from '../Login/Login';
 
 const Splash = () => {
   return (
     <>
       <h1>Splash Page</h1>
-      <LoginButton/>
-      <LogoutButton/>
+      <Login/>
       <Profile/>
     </>
   );
